refactor(channels): extract channels feature seam into packages/core/src/channels
Create a dedicated channels feature so future channel-identity unification is a
change to one service's internals rather than to every call site. Pure code
relocation with no behavior change; still gated by the existing project-bluebird
flag via the routes that mount these surfaces.

- Move ChannelTasksService -> ChannelsService (channels/channels.ts) with an
  IChannelsService interface; inject the shared DesktopFsClient from canvas.
- Move channel task schemas -> channels/schemas.ts.
- Move validateChannelName -> channels/channelName.ts.
- Move ChannelLinkService -> channels/channel-link.ts.
- Add channels/identifiers.ts (CHANNELS_SERVICE, CHANNEL_LINK_SERVICE) and
  channels/channels.module.ts (channelsCoreModule), loaded beside canvasCoreModule.
- Rewire host-router routers, apps/code DI, and UI imports; drop the old
  channel bindings/tokens from canvas and links.

No core channel store is introduced: channel data state currently lives in
React Query caches in the UI, so an empty domain store would be a dead
abstraction. The api-client-direct channel/task-channel queries are left on
their existing transport; routing them through ChannelsService is the deferred
client migration this seam is designed to absorb.
