+++ packages/core/src/links/identifiers.ts
@@ -21,6 +21,3 @@ export const OPEN_TARGET_LINK_SERVICE = Symbol.for(
   "posthog.core.openTargetLinkService",
 );
 export const CANVAS_LINK_SERVICE = Symbol.for("posthog.core.canvasLinkService");
-export const CHANNEL_LINK_SERVICE = Symbol.for(
-  "posthog.core.channelLinkService",
-);
