+++ packages/host-router/src/routers/channel-tasks.router.ts
@@ -1,11 +1,11 @@
+import type { IChannelsService } from "@posthog/core/channels/channels";
+import { CHANNELS_SERVICE } from "@posthog/core/channels/identifiers";
 import {
   channelTaskIdInput,
   channelTaskRecordSchema,
   fileChannelTaskInput,
   listChannelTasksInput,
-} from "@posthog/core/canvas/channelTaskSchemas";
-import { CHANNEL_TASKS_SERVICE } from "@posthog/core/canvas/identifiers";
-import type { IChannelTasksService } from "@posthog/core/canvas/services";
+} from "@posthog/core/channels/schemas";
 import { publicProcedure, router } from "@posthog/host-trpc/trpc";
 import { z } from "zod";
 
@@ -15,22 +15,18 @@ export const channelTasksRouter = router({
     .output(z.array(channelTaskRecordSchema))
     .query(({ ctx, input }) =>
       ctx.container
-        .get<IChannelTasksService>(CHANNEL_TASKS_SERVICE)
+        .get<IChannelsService>(CHANNELS_SERVICE)
         .list(input.channelId),
     ),
   file: publicProcedure
     .input(fileChannelTaskInput)
     .output(channelTaskRecordSchema)
     .mutation(({ ctx, input }) =>
-      ctx.container
-        .get<IChannelTasksService>(CHANNEL_TASKS_SERVICE)
-        .file(input),
+      ctx.container.get<IChannelsService>(CHANNELS_SERVICE).file(input),
     ),
   unfile: publicProcedure
     .input(channelTaskIdInput)
     .mutation(({ ctx, input }) =>
-      ctx.container
-        .get<IChannelTasksService>(CHANNEL_TASKS_SERVICE)
-        .unfile(input.id),
+      ctx.container.get<IChannelsService>(CHANNELS_SERVICE).unfile(input.id),
     ),
 });
